requirements.txt
Django==3.2.13
djangorestframework==3.13.1
celery==5.2.6
Deprecated==1.2.13
PyJWT==2.3.0
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
django-celery-beat==2.2.1
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework-simplejwt==5.1.0
kombu==5.2.4
packaging==21.3
pip==22.0.4
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
setuptools==62.2.0
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.1
psycopg2==2.9.3
flake8==4.0.1
pre-commit==2.19.0
flake8-annotations==2.9.0
black==22.3.0
